Move cell editing related methods to CellInput
